Move parser to dedicated submodule and make AMQPBaseActor non parser-specific
